make API used by modules available externally
- use case: modules that are not compiled directly into Darshan but
would like to register modules and store data there
